+++ src/main/java/net/grid/vampiresdelight/common/mixin/MixinHunterCampPieces.java
@@ -27,7 +27,7 @@ public class MixinHunterCampPieces {
     @Shadow
     private boolean advanced;
 
-    @Inject(at = @At("TAIL"), method = "postProcess(Lnet/minecraft/world/level/WorldGenLevel;Lnet/minecraft/world/level/StructureManager;Lnet/minecraft/world/level/chunk/ChunkGenerator;Lnet/minecraft/util/RandomSource;Lnet/minecraft/world/level/levelgen/structure/BoundingBox;Lnet/minecraft/world/level/ChunkPos;Lnet/minecraft/core/BlockPos;)V", remap = false)
+    @Inject(at = @At("TAIL"), method = "postProcess(Lnet/minecraft/world/level/WorldGenLevel;Lnet/minecraft/world/level/StructureManager;Lnet/minecraft/world/level/chunk/ChunkGenerator;Lnet/minecraft/util/RandomSource;Lnet/minecraft/world/level/levelgen/structure/BoundingBox;Lnet/minecraft/world/level/ChunkPos;Lnet/minecraft/core/BlockPos;)V")
     public void placeCookingPot(WorldGenLevel worldIn, StructureManager structureManager, ChunkGenerator chunkGenerator, RandomSource random, BoundingBox structureBoundingBoxIn, ChunkPos chunkPos, BlockPos blockPos, CallbackInfo ci) {
         if (random.nextInt(100) <= VDConfiguration.COOKING_POT_IN_HUNTER_CAMP_CHANCE.get() && VDConfiguration.GENERATE_COOKING_POT_IN_HUNTER_CAMP.get()) {
             ((StructurePieceAccessor) this).vampiresdelight$placeBlock(worldIn, Blocks.CAMPFIRE.defaultBlockState(), advanced ? 1 : 3, 0, advanced ? 3 : 1, structureBoundingBoxIn);